[WIP] buildable (still have lint issues)
