+++ handler.go
@@ -6,9 +6,9 @@ import (
 	"net/url"
 	"os"
 
-	"github.com/hirakiuc/gonta-app/event"
 	"github.com/hirakiuc/gonta-app/handler"
 	"github.com/hirakiuc/gonta-app/log"
+	"github.com/hirakiuc/gonta-app/parser"
 	"go.uber.org/zap"
 )
 
@@ -37,13 +37,13 @@ func Serve(w http.ResponseWriter, r *http.Request) {
 		return
 	}
 
-	err = handler.HandleEvent(w, result.type, result.JSON, log)
+	err = handleEvent(w, result, log)
 	if err != nil {
 		return
 	}
 }
 
-func parseBody(w http.ResponseWriter, r *http.Request, log *zap.Logger) (*event.BodyParseResult, error) {
+func parseBody(w http.ResponseWriter, r *http.Request, log *zap.Logger) (*parser.BodyParseResult, error) {
 	buf, err := ioutil.ReadAll(r.Body)
 	if err != nil {
 		log.Error("Failed to read request body", zap.Error(err))
@@ -62,9 +62,9 @@ func parseBody(w http.ResponseWriter, r *http.Request, log *zap.Logger) (*event.
 
 	log.Debug("Received body", zap.String("body", jsonStr))
 
-	parser := event.NewBodyParser()
+	bodyParser := parser.NewBodyParser()
 
-	result, err := parser.Parse(jsonStr)
+	result, err := bodyParser.Parse(jsonStr)
 	if err != nil {
 		log.Error("Failed to parse event", zap.Error(err))
 		w.WriteHeader(http.StatusInternalServerError)
@@ -77,4 +77,36 @@ func parseBody(w http.ResponseWriter, r *http.Request, log *zap.Logger) (*event.
 
 func getVerificationToken() string {
 	return os.Getenv("VERIFICATION_TOKEN")
+}
+
+func handleEvent(w http.ResponseWriter, result *parser.BodyParseResult, log *zap.Logger) error {
+	eventParser := parser.NewEventParser()
+
+	switch result.Type {
+	case "url_verification":
+		e, err := eventParser.ParseURLVerificationEvent(result.JSON)
+		if err != nil {
+			log.Error("Failed to parse the URLVerificationEvent", zap.Error(err))
+			w.WriteHeader(http.StatusInternalServerError)
+
+			return err
+		}
+
+		handler := handler.NewURLVerificationHandler()
+		return handler.Handle(w, e)
+
+	default:
+		e, err := eventParser.ParseCallbackEvent(result.JSON)
+		if err != nil {
+			log.Error("Failed to parse the CallbackEvent", zap.Error(err))
+			w.WriteHeader(http.StatusInternalServerError)
+
+			return err
+		}
+
+		handler := handler.NewMentionHandler()
+		return handler.Handle(w, e)
+	}
+
+	return nil
 }
