+++ handler/empty_handler.go
@@ -3,7 +3,7 @@ package handler
 import (
 	"net/http"
 
-	event "github.com/hirakiuc/gonta-app/event"
+	"github.com/hirakiuc/gonta-app/model"
 )
 
 // EmptyHandler describe a replyer with empty response.
@@ -15,7 +15,7 @@ func NewEmptyHandler() *EmptyHandler {
 }
 
 // Reply respond an empty response.
-func (r *EmptyHandler) Reply(w http.ResponseWriter, msg *event.CallbackEvent) error {
+func (r *EmptyHandler) Reply(w http.ResponseWriter, msg *model.CallbackEvent) error {
 	w.WriteHeader(http.StatusOK)
 
 	return nil
