+++ handler/mention_handler.go
@@ -3,8 +3,8 @@ package handler
 import (
 	"net/http"
 
-	"github.com/hirakiuc/gonta-app/event"
 	"github.com/hirakiuc/gonta-app/log"
+	"github.com/hirakiuc/gonta-app/model"
 )
 
 // MentionHandler describe a instance of MentionHandler.
@@ -16,7 +16,7 @@ func NewMentionHandler() *MentionHandler {
 }
 
 // Reply send a response to the slack.
-func (req *MentionHandler) Handle(w http.ResponseWriter, msg *event.CallbackEvent) error {
+func (req *MentionHandler) Handle(w http.ResponseWriter, msg *model.CallbackEvent) error {
 	logger := log.GetLogger()
 	logger.Debug("MentionHandler handle")
 	w.WriteHeader(http.StatusOK)
